Reflection: Type-cast bugfix
Apparently base-class also as CustomAttrinutes, but these are of of
different type. Using types GUID from an instance known to have the
correct type before the cast.

Note: This is a hack. I'm sure there are other nicer ways to determine
if an object is an instance of a class.